requirements/requirements-clean.txt
# InfraSpectre Production Requirements
# Generated: September 11, 2025
# Python 3.11+ compatible

# Core Dependencies
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.28.2
PyYAML==6.0.2
cryptography==43.0.1
pycryptodome==3.21.0

# Data & Storage
sqlite3  # Built-in to Python
prometheus_client==0.21.0

# Testing & Development (optional for production)
pytest==8.4.1
pytest-asyncio==0.24.0
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Observability
prometheus-client==0.21.0
